move hatch locator
